feat: renaming some of the public interfaces, changes on unit tests
