+++ Sources/PrimerSDK/Classes/Core/PrimerHeadlessUniversalCheckout/Composable/FormWithRedirect/Ideal/BanksCollectableData.swift
@@ -1,0 +1,12 @@
+//
+//  BanksCollectableData.swift
+//  PrimerSDK
+//
+//  Created by Alexandra Lovin on 21.11.2023.
+//
+
+import Foundation
+public enum BanksCollectableData: FormCollectableData {
+    case bankId(bankId: String)
+    case bankFilterText(text: String)
+}
