+++ Sources/PrimerSDK/Classes/Core/PrimerHeadlessUniversalCheckout/Composable/FormWithRedirect/Ideal/BanksComponent.swift
@@ -1,162 +1,9 @@
 //
-//  IdealComponent.swift
+//  BanksComponent.swift
 //  PrimerSDK
 //
-//  Created by Alexandra Lovin on 06.11.2023.
+//  Created by Alexandra Lovin on 21.11.2023.
 //
 
 import Foundation
-
-public enum BanksCollectableData: FormCollectableData {
-    case bankId(bankId: String)
-    case bankFilterText(text: String)
-}
-
-public class BanksComponent: PrimerHeadlessFormComponent {
-
-    public typealias T = BanksCollectableData
-    let paymentMethodType: PrimerPaymentMethodType
-    
-    public weak var errorDelegate: PrimerHeadlessErrorableDelegate?
-    public weak var validationDelegate: PrimerHeadlessValidatableDelegate?
-    public weak var stepDelegate: PrimerHeadlessSteppableDelegate?
-
-    public private(set) var nextDataStep: BanksStep = .loading
-    private(set) var banks: [IssuingBank] = []
-    private(set) var bankId: String?
-    private let createWebRedirectComponent: () -> WebRedirectComponent
-    private let onFinished: (Bool) -> (Void)
-    private let tokenizationModelDelegate: BankSelectorTokenizationDelegate
-
-    init(paymentMethodType: PrimerPaymentMethodType, tokenizationModelDelegate: BankSelectorTokenizationDelegate, createWebRedirectComponent: @escaping () -> WebRedirectComponent, onFinished: @escaping (Bool) -> (Void)) {
-        self.paymentMethodType = paymentMethodType
-        self.tokenizationModelDelegate = tokenizationModelDelegate
-        self.createWebRedirectComponent = createWebRedirectComponent
-        self.onFinished = onFinished
-    }
-    
-    public func updateCollectedData(collectableData: BanksCollectableData) {
-        validateData(for: collectableData)
-        switch collectableData {
-        case .bankId(bankId: let bankId):
-            self.bankId = bankId
-            if isBankIdValid(bankId: bankId) {
-                // todo: here need to figure the tokenization workflow and more refactoring
-                tokenizationModelDelegate.tokenize(bankId: bankId)
-                let redirectComponent = createWebRedirectComponent()
-               // notify
-            }
-        case .bankFilterText(text: let text):
-            let filteredBanks = tokenizationModelDelegate.filterBanks(query: text)
-            stepDelegate?.didReceiveStep(step: BanksStep.banksRetrieved(banks: filteredBanks.map { IssuingBank(bank: $0) }))
-        }
-    }
-
-    func isBankIdValid(bankId: String) -> Bool {
-        banks.compactMap({ $0.id }).contains(bankId)
-    }
-    
-    func validateData(for data: BanksCollectableData) {
-        validationDelegate?.didUpdate(validationStatus: .validating, for: data)
-        switch data {
-        case .bankId(bankId: let bankId):
-            if !isBankIdValid(bankId: bankId) {
-                let userInfo = [
-                    "file": #file,
-                    "class": "\(Self.self)",
-                    "function": #function,
-                    "line": "\(#line)"
-                ]
-                let error = banks.isEmpty ? PrimerValidationError.banksNotLoaded(
-                    userInfo: userInfo,
-                    diagnosticsId: UUID().uuidString) :
-                    PrimerValidationError.invalidBankId(
-                    bankId: bankId,
-                    userInfo: userInfo,
-                    diagnosticsId: UUID().uuidString)
-                ErrorHandler.handle(error: error)
-                validationDelegate?.didUpdate(validationStatus: .invalid(errors: [error]), for: data)
-            } else {
-                validationDelegate?.didUpdate(validationStatus: .valid, for: data)
-            }
-        case .bankFilterText(text: _):
-            if banks.isEmpty {
-                let error = PrimerValidationError.banksNotLoaded(
-                    userInfo: [
-                        "file": #file,
-                        "class": "\(Self.self)",
-                        "function": #function,
-                        "line": "\(#line)"
-                    ],
-                    diagnosticsId: UUID().uuidString)
-                ErrorHandler.handle(error: error)
-                validationDelegate?.didUpdate(validationStatus: .invalid(errors: [error]), for: data)
-            } else {
-                validationDelegate?.didUpdate(validationStatus: .valid, for: data)
-            }
-        }
-    }
-
-    public func start() {
-        trackStart()
-        stepDelegate?.didReceiveStep(step: BanksStep.loading)
-        tokenizationModelDelegate.retrieveListOfBanks()
-            .done { banks -> Void in
-                self.banks = banks.map { IssuingBank(bank: $0) }
-                self.stepDelegate?.didReceiveStep(step: BanksStep.banksRetrieved(banks: self.banks))
-            }.catch { errror in
-                print("Error")
-            }
-    }
-    
-    public func submit() {
-        trackSubmit()
-        switch nextDataStep {
-        case .loading: break
-        case .banksRetrieved(banks: _): break
-        }
-    }
-
-    public func cancel() {
-        tokenizationModelDelegate.cancel()
-    }
-}
-
-private extension BanksComponent {
-    func trackSubmit() {
-        trackEvent(BanksAnalyticsEvent.submit)
-    }
-    func trackStart() {
-        trackEvent(BanksAnalyticsEvent.start)
-    }
-    func trackCollectableData() {
-        trackEvent(BanksAnalyticsEvent.updateCollectedData)
-    }
-    func trackEvent(_ event: BanksAnalyticsEvent, additionalParams: [String: String]? = nil) {
-        var params: [String: String] = ["category": paymentMethodType.rawValue]
-        if let additionalParams {
-            params.merge(additionalParams) { (_, new) in new }
-        }
-        let sdkEvent = Analytics.Event(
-            eventType: .sdkEvent,
-            properties: SDKEventProperties(
-                name: event.rawValue,
-                params: params))
-        Analytics.Service.record(events: [sdkEvent])
-    }
-}
-
-public extension BanksComponent {
-    @objc final class IssuingBank: NSObject {
-        public let id: String
-        public let name: String
-        public let iconUrlStr: String?
-        public let isDisabled: Bool
-        init(bank: AdyenBank) {
-            self.id = bank.id
-            self.name = bank.name
-            self.iconUrlStr = bank.iconUrlStr
-            self.isDisabled = bank.disabled
-        }
-    }
-}
+public protocol BanksComponent: PrimerHeadlessFormComponent where T == BanksCollectableData, P == BanksStep { }
