+++ Sources/PrimerSDK/Classes/Core/PrimerHeadlessUniversalCheckout/Managers/Payment Method Managers/PrimerHeadlessUniversalCheckoutComponentWithRedirectManager.swift
@@ -7,25 +7,19 @@
 
 import Foundation
 extension PrimerHeadlessUniversalCheckout {
-    public class PrimerHeadlessFormWithRedirectManager: NSObject {
-        let paymentMethodType: PrimerPaymentMethodType
-        public init?(paymentMethodType: String) {
+    public class ComponentWithRedirectManager: NSObject {
+        public func provide(paymentMethodType: String) -> (any BanksComponent)? {
             guard let paymentMethodType = PrimerPaymentMethodType(rawValue: paymentMethodType) else {
                 return nil
             }
             guard paymentMethodType == .adyenIDeal else {
                 return nil
             }
-            self.paymentMethodType = paymentMethodType
-        }
-        public func provideBanksComponent() -> BanksComponent? {
             guard let tokenizationModelDelegate = PrimerAPIConfiguration.paymentMethodConfigViewModels.first(where: { $0 is BankSelectorTokenizationDelegate }) as? BankSelectorTokenizationDelegate  else {
                 return nil
             }
-            return BanksComponent(paymentMethodType: paymentMethodType, tokenizationModelDelegate: tokenizationModelDelegate, createWebRedirectComponent: {
-                WebRedirectComponent(paymentMethodType: self.paymentMethodType, tokenizationModelDelegate: tokenizationModelDelegate)
-            }) { result in
-                print("Did finish with result")
+            return DefaultBanksComponent(paymentMethodType: paymentMethodType, tokenizationModelDelegate: tokenizationModelDelegate) {
+                WebRedirectComponent(paymentMethodType: paymentMethodType, tokenizationModelDelegate: tokenizationModelDelegate)
             }
         }
     }
